fix: remove redundant u32 casts on mozjpeg_sys JDIMENSION fields
JDIMENSION is u32, so `ci.output_width as u32` is unnecessary.
Triggers clippy::unnecessary_cast with -D warnings in CI.

+++ zenjpeg/tests/dump_decoder_diffs.rs
@@ -34,11 +34,11 @@ fn decode_mozjpeg(data: &[u8]) -> Option<(u32, u32, Vec<u8>)> {
         }
         ci.out_color_space = J_COLOR_SPACE::JCS_RGB;
         jpeg_start_decompress(&mut ci);
-        let w = ci.output_width as u32;
-        let h = ci.output_height as u32;
+        let w = ci.output_width;
+        let h = ci.output_height;
         let stride = w as usize * ci.output_components as usize;
         let mut out = vec![0u8; h as usize * stride];
-        while (ci.output_scanline as u32) < h {
+        while ci.output_scanline < h {
             let off = ci.output_scanline as usize * stride;
             let mut p = out[off..].as_mut_ptr();
             jpeg_read_scanlines(&mut ci, &mut p, 1);
